Keep composer send button in view at all times
When composing long emails, or emails with a long quoted thread inside
of it, the Send button could end up scrolling down below the viewport,
forcing you to scroll down to the end of the email to send it.

Sticky positioning allows us to keep the Send button in view, when it
would usually have disappeared below the viewport.